requirements.txt
Django>2.2,<=3.2
django-import-export
djangorestframework
djangorestframework-simplejwt
django-admin-list-filter-dropdown
django-rest-swagger>=2.2.0
django-parler
django-parler-rest
django-modeladmin-reorder
django-bootstrap-datepicker-plus
django-smart-selects>=1.5.9
django-storages[azure]
django-admin-menu>=1.5
git+git://github.com/ahoazure/django-data-wizard.git
msal>=1.9.0
mysqlclient>=2.0.1
Pillow>=6.0.0
python-dotenv>=0.17.1
pytz>=2018.5
tablib[all]
whitenoise
